PRSD-358: Moved data model to new folder
